requirements.txt
ahk==0.9.0
astroid==2.3.3
colorama==0.4.3
comtypes==1.1.7
isort==4.3.21
Jinja2==2.11.2
keyboard==0.13.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.0.4
numpy==1.18.3
Pillow==7.2.0
psutil==5.7.0
PyAutoGUI==0.9.47
pygame==1.9.6
PyGetWindow==0.0.7
pylint==2.4.4
PyMsgBox==1.0.8
pynput==1.4.2
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
pywin32==225
pywinauto==0.6.8
selenium==3.141.0
six==1.12.0
typed-ast==1.4.1
urllib3==1.25.6
wrapt==1.11.2